Details unit separate added
